tests moved to a separate folder
