+++ src/tests/app.controller.spec.ts
@@ -1,9 +1,9 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { AppController } from '../app.controller';
 import { AppService } from '../app.service';
-import { BinanceService } from '../exchanges/binance/binance.service';
-import { KrakenService } from '../exchanges/kraken/kraken.service';
-import { HuobiService } from '../exchanges/huobi/huobi.service';
+import { BinanceService } from '../exchanges/services/binance.service';
+import { KrakenService } from '../exchanges/services/kraken.service';
+import { HuobiService } from '../exchanges/services/huobi.service';
 
 describe('AppController', () => {
   let appController: AppController;
