+++ src/tests/app.service.spec.ts
@@ -1,8 +1,8 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { AppService } from './app.service';
-import { BinanceService } from './exchanges/binance/binance.service';
-import { KrakenService } from './exchanges/kraken/kraken.service';
-import { HuobiService } from './exchanges/huobi/huobi.service';
+import { AppService } from '../app.service';
+import { BinanceService } from '../exchanges/services/binance.service';
+import { KrakenService } from '../exchanges/services/kraken.service';
+import { HuobiService } from '../exchanges/services/huobi.service';
 
 describe('AppService', () => {
   let appService: AppService;
